+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - Set the print style with `--print`. Add new print styles: `debug`, `json`.
 - Add manpages. Open them with `uamp man`.
 - In CLI `config` add `--path` as alias to `--print-path`.
+- Option to restart uamp with `restart` control message.
+- Uamp will automatically restart if its binary changes.
 
 ### Changes
 - CLI now uses `{` and `}` instead of `[` and `]`. The old way is still